move util methods out Transaction
